requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# NLP
spacy==3.7.2

# Document Processing
PyMuPDF>=1.24.0
python-docx==1.1.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
reportlab==4.0.9

# Utilities
python-dateutil==2.8.2
